remove CustomerValidationService and its interface
