+++ CarRental/CarRental.Logic/Services/CustomerService.cs
@@ -3,6 +3,8 @@ using CarRental.DAL.Entities;
 using CarRental.DAL.Repositories;
 using CarRental.Logic.Models;
 using CarRental.Logic.Services.IServices;
+using FluentValidation;
+using FluentValidation.Results;
 using Microsoft.Extensions.Logging;
 
 namespace CarRental.Logic.Services;
@@ -12,10 +14,10 @@ public class CustomerService : ICustomerService
     private readonly IRepository<Customer> _customerRepository;
     private readonly IMapper _mapper;
     private readonly ILogger _logger;
-    private readonly ICustomerValidationService _validator;
+    private readonly IValidator<CustomerViewModel> _validator;
 
     public CustomerService(IRepository<Customer> customerRepository, IMapper mapper, ILogger<CustomerService> logger,
-        ICustomerValidationService validator)
+        IValidator<CustomerViewModel> validator)
     {
         _customerRepository = customerRepository;
         _mapper = mapper;
@@ -37,7 +39,7 @@ public class CustomerService : ICustomerService
 
     public void Create(CustomerViewModel model)
     {
-        if (!_validator.IsAllValid(model))
+        if (!IsAllValid(model))
         {
             throw new ArgumentException("Customer is not valid.");
         }
@@ -47,7 +49,7 @@ public class CustomerService : ICustomerService
 
     public void Update(CustomerViewModel model)
     {
-        if (!_validator.IsAllValid(model))
+        if (IsAllValid(model))
         {
             throw new ArgumentException("Customer is not valid.");
         }
@@ -61,4 +63,21 @@ public class CustomerService : ICustomerService
         _customerRepository.Delete(id);
         _logger.LogInformation($"Customer with id {id} was deleted.");
     }
+
+    #region Validation
+    public bool IsAllValid(CustomerViewModel customer)
+    {
+        var validationResult = _validator.Validate(customer);
+        LogErrors(validationResult);
+        return validationResult.IsValid;
+    }
+
+    private void LogErrors(ValidationResult validationResult)
+    {
+        foreach (var failure in validationResult.Errors)
+        {
+            _logger.LogInformation("Property " + failure.PropertyName + " failed validation. Error was: " + failure.ErrorMessage);
+        }
+    }
+    #endregion Validation
 }
