--- CarRental/CarRental.Logic/Services/IServices/ICustomerValidationService.cs
@@ -1,7 +1,0 @@
-﻿using CarRental.Logic.Models;
-
-namespace CarRental.Logic.Services.IServices;
-public interface ICustomerValidationService
-{
-    public bool IsAllValid(CustomerViewModel customer);
-}
